# Improve fitting performance of most non-linear fitters