--- a/docs/changes/modeling/16673.perf.rst
+++ b/docs/changes/modeling/16673.perf.rst
@@ -1,0 +1,1 @@
+Performance of most non-linear fitters has been significantly improved by reducing the overhead in evaluating models inside the objective function.