@astrofrog 2024-07-04
I've split out some of the changes here that relate to the Model API into https://github.com/astropy/astropy/pull/16677