@pllim 2024-07-11
Please use the new `perf` change log category, see:

* https://github.com/astropy/astropy/pull/16694