@astrofrog 2024-07-15
The asv benchmarks seem to be working! The one regression is a fluke.