@astrofrog 2024-07-15
For convenience:

´´´

2024-07-11T16:08:05.5964106Z | Change   | Before [ff59e9cf]    | After [2a05840e]    |   Ratio | Benchmark (Parameter)                                         |
2024-07-11T16:08:05.5966580Z |----------|----------------------|---------------------|---------|---------------------------------------------------------------|
2024-07-11T16:08:05.5968188Z | -        | 103±3ms              | 63.2±0.3ms          |    0.61 | modeling.fitting.time_datafit_compound_LevMarLSQFitter        |
2024-07-11T16:08:05.5972159Z | -        | 2.17±0.03ms          | 968±7μs             |    0.45 | modeling.fitting.time_Gaussian1D_LevMarLSQFitter              |
2024-07-11T16:08:05.5973944Z | -        | 368±60μs             | 162±3μs             |    0.44 | units.time_quantity_array_conversion                          |
2024-07-11T16:08:05.5975459Z | -        | 289±70μs             | 92.4±0.7μs          |    0.32 | units.time_quantity_init_array                                |
2024-07-11T16:08:05.5977071Z | -        | 260±2ms              | 64.9±0.3ms          |    0.25 | modeling.fitting.time_large_gauss_combined_1d_LevMarLSQFitter |
2024-07-11T16:08:05.5978602Z | -        | 19.9±4ms             | 4.22±0.02ms         |    0.21 | modeling.fitting.time_combined_gauss_1d_LevMarLSQFitter       |
2024-07-11T16:08:05.5980178Z | -        | 176±30μs             | 35.8±0.2μs          |    0.2  | units.time_quantity_creation                                  |
2024-07-11T16:08:05.5981557Z | Change   | Before [ff59e9cf]    | After [2a05840e]    |   Ratio | Benchmark (Parameter)          |
2024-07-11T16:08:05.5982920Z |----------|----------------------|---------------------|---------|--------------------------------|
2024-07-11T16:08:05.5984049Z | +        | 41.8±0.1μs           | 83.2±40μs           |    1.99 | units.time_quantity_times_unit |
2024-07-11T16:08:05.5985127Z 
2024-07-11T16:08:05.5985548Z SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
´´´