@pllim 2024-08-10
Does this comment need to be resolved? 👉 https://github.com/astropy/astropy/pull/16673#issuecomment-2228888779