@astrofrog 2024-08-10
No I was just commenting that the units asv tests were unreliable but not related to the changes here