@astrofrog 2024-09-10
I have an idea to fix this, will take a look this morning