@neutrinoceros 2024-09-10
Thanks for the heads up @astrofrog ! I was about to give it a try myself, glad we could avoid another race condition :)